moving timer implementation so that people that dont use newrelic don't have to import it
